chore(ui): rename tag group to tag toggle group to reduce confusion
