--- src/common/components/TagGroup/index.ts
@@ -1,1 +1,0 @@
-export * from "./TagGroup";
